Move constants into IVC crate
