+++ ivc/src/poseidon/params.rs
@@ -1,5 +1,5 @@
 use ark_bn254::Fr as Fp;
-use mina_poseidon::poseidon::ArithmeticSpongeParams;
+use mina_poseidon::{constants::SpongeConstants, poseidon::ArithmeticSpongeParams};
 use once_cell::sync::Lazy;
 
 // FIXME: move into mina_poseidon when this code is ready to go into production.
@@ -837,4 +837,23 @@ fn params() -> ArithmeticSpongeParams<Fp> {
 pub fn static_params() -> &'static ArithmeticSpongeParams<Fp> {
     static PARAMS: Lazy<ArithmeticSpongeParams<Fp>> = Lazy::new(params);
     &PARAMS
+}
+
+/// Constants used by the IVC circuit used by the folding scheme
+/// This is meant to be only used for the IVC circuit, with the BN254 curve
+/// It has not been tested with/for other curves.
+// This must be moved into mina_poseidon later.
+#[derive(Clone)]
+pub struct PlonkSpongeConstantsIVC {}
+
+impl SpongeConstants for PlonkSpongeConstantsIVC {
+    const SPONGE_CAPACITY: usize = 1;
+    const SPONGE_WIDTH: usize = 3;
+    const SPONGE_RATE: usize = 2;
+    const PERM_ROUNDS_FULL: usize = 55;
+    const PERM_ROUNDS_PARTIAL: usize = 0;
+    const PERM_HALF_ROUNDS_FULL: usize = 0;
+    const PERM_SBOX: u32 = 7;
+    const PERM_FULL_MDS: bool = true;
+    const PERM_INITIAL_ARK: bool = false;
 }
